Refactor and reorganize utility imports and logic
Moved various utility functions to specialized packages for better modularity and readability. Updated device filtering, version parsing, and PDF generation logic to use the newly organized packages. Simplified device registration flow and error handling.
